chore: move CustomizeSearch to customizesearch folder
